Refactor added Attrio::Collections::Common changed collections to use it
