+++ lib/attrio/collections/set.rb
@@ -1,9 +1,10 @@
-
 module Attrio
   module Collections
     class Set < DelegateClass(Set)
+      include Attrio::Collections::Common
 
-      def initialize
+      def initialize(type, options)
+        @type       = type, @options = options
         @collection = ::Set.new
         super(@collection)
       end
